feat: 회원 포인트 조회 어댑터 구현(#114)

+++ src/main/java/com/example/green/domain/mypage/client/PointTotalGetClient.java
@@ -1,5 +1,7 @@
 package com.example.green.domain.mypage.client;
 
+import java.math.BigDecimal;
+
 public interface PointTotalGetClient {
 
 	/**
@@ -8,5 +10,5 @@ public interface PointTotalGetClient {
 	 * @param userId 사용자 ID
 	 * @return 포인트 정보
 	 */
-	int getTotalPoints(Long userId);
+	BigDecimal getTotalPoints(Long userId);
 }
